DIPviewer interface throws error if we're trying to show a 0D image.
Python crashed hard when passing a 0D image to `Show`. An exception
is thrown somewhere deep in the bowels of DIPviewer, running in a
different thread. That exception causes `abort()` to be called, since
you cannot catch exceptions thrown in a different thread. This way
we throw the exception in the interface, where it can be caught normally.