Moved unit tests out of header files and into source files.
Unit tests sometimes need to include additional headers, which causes
the set of headers loaded within a compilation unit to change if unit
tests are disabled. This was annoying.

It was also annoying having to recompile so many files when making a
change to one of the unit tests.

It is nice having the unit tests close to the sources, even for
header-only functionality. But I've come to the conclusion that it's
more annoying that nice.