Replace global static by function wrapping a static variable
This fixes a really strange crash on shutdown in our tst_DipImage.py
test, which seems to crash when trying to destruct the featureIndices_
map on shutdown.

Generally, global statics are usually a bad idea, but quite frankly
I have no idea what exactly is happening here. It may be that this
patch just hides the issue and it would pop up again when the
tool would be actually used via the python API now...