Allow building against system libs instead of bundled dependencies
When DIP_BUILD_BUNDLED_DEPENDENCIES=OFF is passed to cmake, we
will rely on find_package to find Eigen3, zlib, jpeg, ics and tiff
libraries. The fftw_api.h header will always be used, as it's not
public API that's included in the fftw3 package on ArchLinux e.g.

The default for this option is still ON, i.e. this is an opt-in
feature.

To allow building against an externally provided doctest, the
includes are normalized to follow the suggested upstream format:

```
#include <doctest/doctest.h>
```

Change-Id: I4ed9095eefbfd889642d1c3d42b3d9af0dfc15ae