Add missing defaulted ctor, dtor and operator= for Histogram
The histogram class has many constructors, but was missing all
of the default constructors. This made it much harder to use, one
would not be able to cmopile `std::vector<dip::Histogram>(5)` e.g.

Personally I don't see a reason why it should be forbidden to
construct an empty histogram.

Change-Id: I0d5e91dcc00a89b8e7a6d00a1d828845a7985cb1