Removed dip::Feature::Base::Scale() again.
This commit undoes some changes introduced in 8c586648.
I don't like making backwards-incompatible changes, but this particular
change hadn't made it to a release yet, so this shouldn't affect people
that cautiously stick to released versions.

dip::Feature::Base::Scale() was an elegant solution, but it introduced some
overhead for calling a virtual function for each feature and each object.
I've solved the issue with a few composite features that combine lengths and
areas (or surface areas and volumes) getting values in incompatible units by
detecting the situation (anisotropic pixels), and explicitly scaling the
area/volume to make the units compatible.

I think this new solution is even neater.