Will this fix GitHub CI on Windows?
CMake on Windows is now identifying `C:/Strawberry/c/lib/libglut.a` as the
FreeGLUT library, but MSVC can't link to it because it's a GCC static library.
This Strawberry thing apparently is a full install of MinGW as part of some
Perl installation: https://github.com/actions/runner-images/issues/5459

In the CI test workflow we download a pre-compiled GLFW library and link to
it. CMake also finds that FreeGLUT library, and so DIPviewer is linked to both.
This should be fine if the FreeGLUT library were compatible with MSVC.

In the deploy workflow we download FreeGLUT sources, compile them, and link
against the result. This should still work. But I don't know why the two
workflows do different things, we should unify them.