Don't import IPython if it's not already imported.
The IPython integration only makes sense in an actual IPython shell,
which implies that IPython has already been imported previously (i.e.,
is in sys.modules).

Previously, IPython would always be imported as long as it is installed,
but that can be quite slow (adding ~0.4s of overhead for me) which is
significant in short-running scripts (that may e.g. be repeatedly
spawned from bash).