Removed the "erika.tif" example image. Updated examples and defaults. [skip ci]
I've been telling people for years to stop using the Lena image, and
all that time I've been fooling myself to believe that the Erika image
is not the same thing. But it is. If we truly want to create a
welcoming environment for everyone, we need to rid it of stuff that
has the potential to make someone unconfortable. There are plenty of
images without sexy poses, let's use those instead.