Fix potential issue in 3D tiff test
Fix an issue in the tiff reading and writing test. In this test, only
the first 5 slices were tested for the 3D images. This worked, as the
test image only had 5 slices, but it wouldn't have caught an accidental
additional slice that was written. This fix addresses that issue, and
ensures all slices of the image are loaded from disk.